+++ src/Convert.hs
@@ -51,6 +51,7 @@ import qualified Convert.StringParam
 import qualified Convert.StringType
 import qualified Convert.Struct
 import qualified Convert.StructConst
+import qualified Convert.SeverityTasks
 import qualified Convert.TFBlock
 import qualified Convert.Typedef
 import qualified Convert.TypeOf
@@ -121,6 +122,7 @@ initialPhases tops selectExclude =
     , Convert.Interface.disambiguate
     , Convert.Package.convert
     , Convert.StructConst.convert
+    , selectExclude Job.SeverityTasks Convert.SeverityTasks.convert
     , Convert.PortDecl.convert
     , Convert.ParamNoDefault.convert tops
     , Convert.ResolveBindings.convert